Move utility classes to a separate package
