+++ app/src/main/java/com/kamron/pogoiv/utils/CommonLogger.java
@@ -1,4 +1,4 @@
-package com.kamron.pogoiv;
+package com.kamron.pogoiv.utils;
 
 import android.util.Log;
