+++ app/src/main/java/com/kamron/pogoiv/utils/DevMethods.java
@@ -1,4 +1,4 @@
-package com.kamron.pogoiv;
+package com.kamron.pogoiv.utils;
 
 import android.graphics.Bitmap;
 import android.os.Environment;
