+++ app/src/main/java/com/kamron/pogoiv/widgets/PokemonSpinnerAdapter.java
@@ -7,7 +7,7 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.TextView;
 
-import com.kamron.pogoiv.GuiUtil;
+import com.kamron.pogoiv.utils.GuiUtil;
 import com.kamron.pogoiv.scanlogic.Pokemon;
 
 import java.util.ArrayList;
